+++ tests/test_elaborator.v
@@ -1,8 +1,10 @@
+From unreleased Extra Dependency "elpi_elaborator.elpi" as elab.
+
 From elpi Require Import elpi.
 
 Elpi Command test.refiner.
 
-Elpi Accumulate File "elpi-elaborator.elpi" From unreleased.
+Elpi Accumulate File elab.
 
 
 Elpi Bound Steps 10000.